% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy.gmt1.R
\docType{data}
\name{gmt.file1}
\alias{gmt.file1}
\title{This is a gmt.file that is from HGNC that should be correct
gmt.file that is used in gene set enrichment analysis that is goo}
\format{A data frame with 3 rows and 1217 variables
\descrinbe{
\item{gmt.file}{List of Gene Familes genesets}
\item{list[1217]}{list of characters}
\item{List[1217]}{List of genes belonging to the gene family}}}
\usage{
gmt.file1
}
\description{
This is a gmt.file that is from HGNC that should be correct
gmt.file that is used in gene set enrichment analysis that is goo
}
\keyword{datasets}
